# apps/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KduFoot API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/exercises": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Search for and filter exercises",
        "description": "Retrieves a paginated list of exercises, with optional filters for keyword search, category, target level, and theme. Requires the 'exercises:read' permission (scope).\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Keyword to find in title or description.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Specific category filter (e.g., Tactical).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "in": "query",
            "description": "Difficulty level filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "theme",
            "in": "query",
            "description": "Specific theme filter (e.g., Finishing).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return.",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip for pagination.",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated search results.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "exercises": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Exercise"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions."
          }
        }
      },
      "post": {
        "tags": [
          "Exercises"
        ],
        "summary": "Create a new exercise drill",
        "description": "Saves a new exercise record to the database. The exercise is automatically associated with the authenticated user. Requires the 'exercises:create' permission.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Details of the exercise to be created.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Exercise"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercise successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "exercise": {
                      "$ref": "#/components/schemas/Exercise"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - User profile not synchronized."
          },
          "403": {
            "description": "Forbidden - Insufficient rights to create exercises."
          },
          "500": {
            "description": "Internal Server Error - Failed to save the record."
          }
        }
      }
    },
    "/api/exercises/{id}": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Retrieve a detailed exercise by ID",
        "description": "Fetches the full details of a single exercise record. Requires the 'exercises:read' permission.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique UUID of the exercise.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved exercise details.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "exercise": {
                      "$ref": "#/components/schemas/Exercise"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Access denied."
          },
          "404": {
            "description": "Not Found - Exercise ID does not exist."
          }
        }
      },
      "put": {
        "tags": [
          "Exercises"
        ],
        "summary": "Update an existing exercise",
        "description": "Modifies the details of an existing exercise. Users can only update exercises they have created. Requires the 'exercises:update' permission.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique ID of the exercise.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Fields to update.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Exercise"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercise successfully updated."
          },
          "403": {
            "description": "Forbidden - Lacks mandatory permissions or not the owner."
          },
          "404": {
            "description": "Not Found - Exercise ID does not exist."
          },
          "500": {
            "description": "Internal Server Error - Update failed."
          }
        }
      },
      "delete": {
        "tags": [
          "Exercises"
        ],
        "summary": "Delete an exercise",
        "description": "Permanently removes an exercise record. Users can only delete their own exercises. Requires the 'exercises:delete' permission.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique ID of the exercise.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful deletion."
          },
          "403": {
            "description": "Forbidden - Unauthorized or not the owner."
          },
          "404": {
            "description": "Not Found - Exercise identifier invalid."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Exercise": {
        "type": "object",
        "description": "A training drill or exercise created by a user.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the exercise."
          },
          "title": {
            "type": "string",
            "description": "Title or name of the exercise."
          },
          "description": {
            "type": "string",
            "description": "Detailed instructions or description of the drill."
          },
          "category": {
            "type": "string",
            "description": "Broad category of the exercise (e.g., Tactical, Technical)."
          },
          "level": {
            "type": "string",
            "description": "Target skill level (e.g., Beginner, Advanced)."
          },
          "theme": {
            "type": "string",
            "description": "Specific theme of the exercise (e.g., Finishing, Passing)."
          },
          "video_url": {
            "type": "string",
            "description": "Optional link to a demonstration video."
          },
          "thumbnail_url": {
            "type": "string",
            "description": "Optional link to an image representing the exercise."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user who created this exercise."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}